+++ src/core/leoGui.ts
@@ -47,7 +47,6 @@ export class LeoGui {
     public frameIndex: number = 0;
 
     public splashScreen: any;
-    public utils: any;
 
     // public ScriptingControllerClass: any; // ? UNEEDED IN LEO-WEB ?
     //
@@ -78,7 +77,6 @@ export class LeoGui {
         this.scriptFileName = undefined;
         this.positionsToAnimate = []; // Redraw cycle should animate these positions and clear this list.
         this.splashScreen = undefined;
-        this.utils = undefined;
 
         // this.ScriptingControllerClass = NullScriptingControllerClass; // ? UNEEDED IN LEO-WEB ?
         //