+++ src/types.ts
@@ -78,6 +78,8 @@ export interface ButtonEntry {
     label?: string; // could have text instead of or in addition to icon, for accessibility and/or if icon fails to load
     tooltip?: string;
     command: string; // Command to execute when the entry is selected
+    visibleFlagsSet?: string[]; // Strings from constants.CONTEXT_FLAGS that need to be true in workspace.getContext() for the entry to be visible
+    visibleFlagsClear?: string[]; // Strings from constants.CONTEXT_FLAGS that need to be false in workspace.getContext() for the entry to be visible
     enabledFlagsSet?: string[]; // Strings from constants.CONTEXT_FLAGS that need to be true in workspace.getContext() for the entry to be enabled
     enabledFlagsClear?: string[]; // Strings from constants.CONTEXT_FLAGS that need to be false in workspace.getContext() for the entry to be enabled
     domElementRef?: HTMLElement; // Reference to the button to refresh its enabled/disabled state.
@@ -320,6 +322,7 @@ export interface LeoPackageStates {
     canHoist: boolean; // Selected node is not the first top node already root
     topIsChapter: boolean; // Top of the hoisted outline is an @chapter node
     hasMarked: boolean; // Has at least one marked node in the outline
+    hasChapters: boolean; // Has at least one @chapter node in the outline
 }
 
 //@+node:felix.20251211000618.10: ** LeoDocument