+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # Change Log
 
+## 1.0.18
+
+- Added select-chapter and select-chapter-main buttons visible only if the outline actually contains @chapter nodes.
+- Fixed cff command as per Leo's #4640
+- Added quick-find buttons to the NAV pane.
+
 ## 1.0.17
 
 - Fixed welcome screen.